Cleanup and correct missing State message fault
The State subkey would be reported as missing for every other subkey that could be present besides State.
While this may never happen the way the evaluation was done is not forwards compatible and is still faulty.

No license is present on this code so strictly speaking no one can use it. A license should be added.

Extra decorative lines at the top were useless and have been removed.

Added `use warnings` as it should be present in any proper Perl code, perhaps using `-w` on schebang line.

Properly converted everything to spaces at 4 character indent for readability. Previously mixed spaces and tabs were used which are inappropriate.

Used early return where reasonable to prevent excess indent and unreadability of code.

Replaced foreach with for everywhere, as foreach hasn't been necessary now for over a decade.

Got rid of scalar(@x)>0 and replaced with proper @x.